Build sveltekit version of infinite wiki
This version has about 90% feature parity. The main missing feature is
linking with breadcrumbs to other wiki pages.
Because svelte does not have functional components that I
can define on the fly in order to inject special render
time functionality, I cannot properly override the markdown
renderer for Links like I can in React.

I may be able to work around this with some kind of context state
that the child link's can read from the article page
but it feels not super worth it when I can stick
with react for the frontend.